+++ NETCORE/CleanPackagesCore.ps1
@@ -1,7 +1,0 @@
-#enable verbose mode
-$VerbosePreference = "Continue";
-
-[String]$dnxRoot = [System.Environment]::ExpandEnvironmentVariables('%USERPROFILE%\.nuget');
-
-rm -r $dnxRoot\packages -Force 
-